I2C Accelerometer interface working
Setup() now puts the accelerometer into active mode before finishing.
The main loop, which is now a switch statement, will fall into a
check_accel functions where I2C is used to interface with accelerometer
to get the x,y, and z acceleration values. Note that these bytes are
being interpretted correctly; the bit 6 is an alert bit and shouldn't be
used in the value.

Some of the other features in the accelerometer need to be explored,
like tap detection, shake detection, and power saving features.