Moved init routines to common.c/h; Accel values corrected
Moved the setup function and separated it into more manageable methods,
they are now in the common.c/h files.

Accelerometer values are handled better in the check_accel() method; ignore
alert bit and determine positive/negative. The values seem quite off, but
I'm sampling at a low rate and I'm not averaging any values. Next up
for the accelerometer is tap and shake detectection. This will require the
setup of port B2 interrupt capabilities, which is partially implemented in
this commit.