Shake Detection working with polling
Getting good shake detection from the accelerometer. However, the pic
interrupt handler and configuration doesnt seem to be working for the
INT2 (PORT B2) external interrupt. The test of the shake detection is
done by polling the INT2 interrupt flag. Therefore, the interrupt is
actually being thrown on the pin and the flag is being set correctly,
but it is not launching into the interrupt vector. Needs more investigation.