Tap detection enabled
Tap detection and shake detection enabled simultaneously. The tap
detection requires the highest sample rate (taking place on the
accelerometer) which will require a larger power draw. This may be
insignificant when compared to the other components.

Still, the larger sample rate makes the shake detection very sensitive.
The tap detection has built in debounce and debounce counters, but it
appears the shake detection does not. Therefore, we may need to debounce
it in software; count a threshold number of shake readings before we
consider it an actual shake.