Tap/Shake Demo with LEDS
Tapping the device (on x,y axis) should turn on green LEDs, shaking the
device will turn off the LEDs. A software debounce is implemented to
make the shaking less sensitive.

The overall sensitivity is a work in progress, the major problem being
that some shakes are also detected as taps, this will need to be worked
 out with some trial and error. The case could potentially help provide
a better "tap vector" too.