Port B interrupt (INT2) working; demo features
The 2nd external interrupt (INT2) is now working properly. This is the
external interrupt attached to the accelerometer that was not throwing
the code into the interrupt vector. The problem: I had moved the vector
location in memory from 0x18 to 0x22. Moving it back to its proper place
(0x18) fixed this issue.

I also added another state, "handle_tilt", to avoid calling the
check_tilt() method inside the interrupt handler. The code takes too
long to run and causes problems in the interrupt itself, leading to some
very odd undefined behavior (program counter jumps all over the place).

The result is that the Buzz+Tap/Shake demo gives more control to the
user as the code no longer has to poll for the interrupt. Functionally,
this means you can still input taps/shakes once the LEDs are all on and
the buzzer is buzzing. Therefore, you can shut off the buzzer by
reseting (shake or tap) or let it play.