Timer 0 setup; A_TONE_* define
Timer 0 setup as a 16 bit timer, with a 1:256 prescaler. Therefore, if
the board is operating at 48MHz, then the timer increments at a rate =
48MHz/256 = 1.875MHz which is 1 count every ~5us. This appears to be a
good starting point for using timer 0 as the piezo driver.

Two defines were added: A_TONE_l and A_TONE_h. These correspond to the
low and high portions of timer 0's starting value. Setting these
differently will result in a different pitch. The low portion is too
insignificant a change to really notice, so alter the high portion to
hear the pitch range. Remember that if the frequency is too low or high
for the piezo, it wont make any sound at all.