Basic song playing ability added
The basic structure for defining and playing songs is exemplified in
this commit (and it works!). A song to be played is defined as an array
of volatile unsigned chars. Each element describes a note (non zero) or
a rest (zero). The timer 0 interrupt simply uses these elements to
determine what to do; the non-rest values correspond to the timer high
value which needs to be set to continually invert the piezo bit. Rests
will just turn the speaker off. The number of times the note has been
inverted is counted which can be checked to determine note length. Once
the note length is reached, the interrupt handler will adjust the
sond_index to move along to the next note in the song (or wrap around if
at the end).

A possible flaw is that different notes of course correspond to
different frequencies of interrupt, which in turn means that the
timer0_counter will represent differing amounts of time. This means that
higher pitched notes could be shorter and lower pitch note could be
played longer if the note length is simply kept constant (as it is now).
This may be imperceivable, or it may not. If it causes a problem, the
note length and counter will have to be scaled depending on the
pitch of the note as the song is played.

Also needed are defines for actual pitch values; A, A#, B, C,C#, etc and
maybe octaves.