Event driven states (or nested states)
A triple buffer, or queue, of events now facilitates the message passing
between states and interrupt vectors.

In order to avoid the problem of interleaving important execution, a
binary sequence pattern is used for operations entering the queue from
within an interrupt vector. However, the interrupt vector must handle the
possibility that the main code did not get the last sequence before it
move to the next. However, in most cases, it can be ignored.

Sequence operation: A global variable seq_num indicates the current
sequence number for the main code. The main code uses this value to
determine which event seq* to enqueue new events from. For instance, if
the seq_num is 1, then seq1 event is enqueued into the event buffer, and
then seq1 is set to the empty event. If an interrupt occurs that needs
to parse messages into the queue (currently only the hardware tilt INT),
then the seq0 (opposite sequence number set) is assigned to the event and
the sequence is toggled to indicate that sequence number (changed to
zero).

This method is a good start, but does not currently have a way to handle
interrupted interrupts, or multiple interrupts attempting to enqueue
messages. This may require multiple sequence pairs with buffers; one pair
for each interrupt vector.