\#pragma code; sequence and event buffer merged
The event buffer struct now contains the sequence events and a pointer
to an event. The pointer points to the current event that the should be
used by the main code. This results in a nice wrapper for these
variables.

Also, the execution still locked up more than once for me. After locking
up with the debugger attached, it appears to be doing the same thing it
was before; jumping seemingly randomly between a few instructions. I
searched around and found that I should have a #pragma code after
defining my interrupt vectors. I haven't been able to lock up the
execution since, but I have tested it that thoroughly.