Added macros for atomic operations
ATOMIC_BEGIN and ATOMIC_END can be used to wrap bodies of code that
should not be interrupted. ATOMIC_BEGIN simply disables interrupts
globally, while ATOMIC_END re-enables them and then checks the hardware
interrupt for the accelerometer. If the flag is set, it will call
check_tilt to read the tilt register.