LED sequences are now recursive-like
The two LED sequences have been changed to fit better with the eventual
USB framework. These functions use static members to keep a persistent
state and repeatedly enqueue their own events until the sequence is
finished. This should allow the sequences to be used within the USB
framework and the process() call.