Synchronous method for int-to-main-loop events
Using two counters, one for the interrupt routine and another for the
main loop, this commit seems to be able to safely transfer events out of
an interrupt vector. This is specifically for the accelerometer's tilt
register, which *must* be read once an interrupt happens.

The counters begin equal, when an interrupt occurs, the interrupt handler
increments its counter. The main loop will compare it's counter with the
interrupt counter on every Get_Next(): if the counters are not equal,
then the main loop will enqueue a tilt event and increment its own
counter.

Additional variables can be used to delay/disable the enqueueing of the
tilt event until the main loop is ready to handle it (EX: main loop
handling a long running execution). Also, the 8-bits of a char is not
really needed, this method could work with a single bit and XOR. If we
end up with several control bits, we could encode them all in a single
char.